requirements.txt
asgiref==3.7.2
black==23.7.0
cli-helpers==2.3.0
click==8.1.5
configobj==5.0.8
Django==4.2.3
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pendulum==2.1.2
pgcli==3.5.0
pgspecial==2.1.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
setproctitle==1.3.2
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.7.1
wcwidth==0.2.6
